Change events storage to unioned hashmaps
Replace EventChance with EventMap, representing the same information but allowing for easy addition and removal of event chances associated with named events.  Include these updates in sett and manager: sett::step now produce an EventMap which can be used by manager::step to calculate event chances on this step.  Add dev mode note in interpreter for future addition of a "developer" mode in console, allowing easier debugging.  Update people to remove Box pointer to the building a hero stays in.